REFACTOR: Split property graphs into own submodule.
