+++ test/categorical_algebra/CSets.jl
@@ -3,7 +3,7 @@ using Test
 
 using Catlab, Catlab.Theories, Catlab.Graphs,
   Catlab.CategoricalAlgebra, Catlab.CategoricalAlgebra.FinSets
-using Catlab.Graphs: TheoryGraph
+using Catlab.Graphs.BasicGraphs: TheoryGraph
 
 # C-set morphisms
 #################
