+++ test/categorical_algebra/StructuredCospans.jl
@@ -3,7 +3,7 @@ using Test
 
 using Catlab, Catlab.Theories, Catlab.Graphs,
   Catlab.CategoricalAlgebra, Catlab.CategoricalAlgebra.FinSets
-using Catlab.Graphs: TheoryGraph
+using Catlab.Graphs.BasicGraphs: TheoryGraph
 
 # Structured cospans of C-sets
 ##############################
